+++ packages/kit/src/runtime/client/start.js
@@ -2,6 +2,7 @@ import { create_client } from './client.js';
 import { init } from './singletons.js';
 import { set_paths } from '../paths.js';
 import { set_public_env } from '../env-public.js';
+import { set_version } from '../env.js';
 
 /**
  * @param {{
@@ -11,13 +12,14 @@ import { set_public_env } from '../env-public.js';
  *     assets: string;
  *     base: string;
  *   },
- *   target: Element;
- *   trailing_slash: import('types').TrailingSlash;
+ *   target: HTMLElement;
+ *   version: string;
  * }} opts
  */
-export async function start({ env, hydrate, paths, target, trailing_slash }) {
+export async function start({ env, hydrate, paths, target, version }) {
 	set_public_env(env);
 	set_paths(paths);
+	set_version(version);
 
 	if (__SVELTEKIT_DEV__ && target === document.body) {
 		console.warn(
@@ -27,8 +29,7 @@ export async function start({ env, hydrate, paths, target, trailing_slash }) {
 
 	const client = create_client({
 		target,
-		base: paths.base,
-		trailing_slash
+		base: paths.base
 	});
 
 	init({ client });