+++ packages/kit/test/apps/options/test/test.js
@@ -151,13 +151,23 @@ test.describe('trailingSlash', () => {
 		expect(await page.textContent('h2')).toBe('/slash/child/');
 	});
 
-	test('ignores trailing slash on endpoint', async ({ baseURL, request }) => {
+	test('removes trailing slash on endpoint', async ({ baseURL, request }) => {
 		const r1 = await request.get('/path-base/endpoint/');
-		expect(r1.url()).toBe(`${baseURL}/path-base/endpoint/`);
+		expect(r1.url()).toBe(`${baseURL}/path-base/endpoint`);
 		expect(await r1.text()).toBe('hi');
 
 		const r2 = await request.get('/path-base/endpoint');
 		expect(r2.url()).toBe(`${baseURL}/path-base/endpoint`);
+		expect(await r2.text()).toBe('hi');
+	});
+
+	test('adds trailing slash to endpoint', async ({ baseURL, request }) => {
+		const r1 = await request.get('/path-base/endpoint-with-slash');
+		expect(r1.url()).toBe(`${baseURL}/path-base/endpoint-with-slash/`);
+		expect(await r1.text()).toBe('hi');
+
+		const r2 = await request.get('/path-base/endpoint-with-slash/');
+		expect(r2.url()).toBe(`${baseURL}/path-base/endpoint-with-slash/`);
 		expect(await r2.text()).toBe('hi');
 	});
 
@@ -167,7 +177,10 @@ test.describe('trailingSlash', () => {
 
 		expect(data).toEqual({
 			type: 'data',
-			nodes: [null, { type: 'data', data: [{ message: 1 }, 'hi'], uses: {} }]
+			nodes: [
+				{ type: 'data', data: [null], uses: {}, slash: 'always' },
+				{ type: 'data', data: [{ message: 1 }, 'hi'], uses: {} }
+			]
 		});
 	});