+++ packages/kit/test/prerendering/trailing-slash/src/service-worker.js
@@ -1,3 +1,0 @@
-import { prerendered } from '$service-worker';
-
-console.log(prerendered);