+++ packages/kit/types/internal.d.ts
@@ -72,6 +72,7 @@ export interface CSRPageNode {
 	component: typeof SvelteComponent;
 	shared: {
 		load?: Load;
+		trailingSlash?: TrailingSlash;
 	};
 	server: boolean;
 }
@@ -157,6 +158,14 @@ export interface Respond {
 	(request: Request, options: SSROptions, state: SSRState): Promise<Response>;
 }
 
+export interface RouteParam {
+	name: string;
+	matcher: string;
+	optional: boolean;
+	rest: boolean;
+	chained: boolean;
+}
+
 /**
  * Represents a route segment in the app. It can either be an intermediate node
  * with only layout/error pages, or a leaf, at which point either `page` and `leaf`
@@ -168,9 +177,7 @@ export interface RouteData {
 
 	segment: string;
 	pattern: RegExp;
-	names: string[];
-	types: string[];
-	optional: boolean[];
+	params: RouteParam[];
 
 	layout: PageNode | null;
 	error: PageNode | null;
@@ -209,6 +216,7 @@ export interface ServerDataNode {
 	type: 'data';
 	data: Record<string, any> | null;
 	uses: Uses;
+	slash?: TrailingSlash;
 }
 
 /**
@@ -266,6 +274,7 @@ export interface SSRNode {
 		prerender?: PrerenderOption;
 		ssr?: boolean;
 		csr?: boolean;
+		trailingSlash?: TrailingSlash;
 	};
 
 	server: {
@@ -273,6 +282,7 @@ export interface SSRNode {
 		prerender?: PrerenderOption;
 		ssr?: boolean;
 		csr?: boolean;
+		trailingSlash?: TrailingSlash;
 		actions?: Actions;
 	};
 
@@ -312,7 +322,7 @@ export interface SSROptions {
 	}): string;
 	app_template_contains_nonce: boolean;
 	error_template({ message, status }: { message: string; status: number }): string;
-	trailing_slash: TrailingSlash;
+	version: string;
 }
 
 export interface SSRErrorPage {
@@ -327,17 +337,14 @@ export interface PageNodeIndexes {
 
 export type SSREndpoint = Partial<Record<HttpMethod, RequestHandler>> & {
 	prerender?: PrerenderOption;
+	trailingSlash?: TrailingSlash;
 };
 
 export interface SSRRoute {
 	id: string;
 	pattern: RegExp;
-	names: string[];
-	types: string[];
-	optional: boolean[];
-
+	params: RouteParam[];
 	page: PageNodeIndexes | null;
-
 	endpoint: (() => Promise<SSREndpoint>) | null;
 }